requirements.txt
Django==3.0.5
psycopg2-binary==2.8.5
Pillow==7.1.1
django-crispy-forms==1.9.0
django-jalali-date
coverage==5.1
selenium==3.141
gunicorn==20.0.4
autopep8==1.5.2
whitenoise==5.0.1
dj-database-url==0.5.0
django-mptt==0.11.0
django-widget-tweaks==1.4.8
django-ckeditor==5.9.0
django-taggit==1.3.0
